refactor: 루트에 엔트리 포인트 생성
