+++ readme.md
@@ -36,6 +36,14 @@
 - [x] Car 관련 검증 로직은 Car 안에서만 처리하기
 - [x] 출력 관련 기능 분리
 
+## 2차 리뷰 반영
+
+- [x] 엔트리 포인트 정의
+- [ ] car static 제거
+- [ ] 랜덤 값 contest에서 받아서 차량 움직이는 방식으로 리팩토링
+- [ ] 확률적 테스트 제거 후, 테스트 보완
+-
+
 ## 고민한 부분
 
 - 랜덤 인풋을 이용하는 부분에서는 어떻게 테스트를 해야할 것인가
