+++ src/main/java/Main.java
@@ -1,0 +1,8 @@
+import application.Application;
+
+public class Main {
+    public static void main(String[] args) {
+        Application application = new Application();
+        application.run();
+    }
+}
